// apps/mobile/src/api/generated/model/expenseCategoryResponseDto.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Fynans API
 * Fynans personal finance API
 * OpenAPI spec version: 1.0
 */

export interface ExpenseCategoryResponseDto {
  id: string;
  /** @nullable */
  parentId: string | null;
  name: string;
  createdAt: string;
  updatedAt: string;
  isConnectedToStore: boolean;
}